generator: js: define custom `console` object
JavaScript, as a scripting language, doesn't define the widely used
`console` object, so we define our own cheap version which only
implements `console.log` and `console.error` to print on the stdout and
stderr respectively.

The actual value-to-string conversion naively uses `toString`, which
gives pretty poor output for compound structures such as objects and
arrays. Users can work around this by using `JSON.stringify`, but this
should probably be implemented in the interface itself.

Enum values are also printed as integers, as that is their actual
representation. Users can work around this by checking the `api.d.ts` to
translate between the integer and enum variants. Improving this output
might be challenging.